requirements.txt
africastalking==1.2.5
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contextlib2==21.6.0
dj-database-url==1.2.0
Django==4.1.7
django-celery-beat==2.4.0
django-filter==22.1
django-heroku==0.3.1
django-phonenumber-field==7.0.2
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gevent==22.10.2
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
phonenumbers==8.13.7
prompt-toolkit==3.0.36
protobuf==4.22.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
python-http-client==3.3.7
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
schema==0.7.5
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
twilio==7.16.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.3.0
zope.event==4.6
zope.interface==6.0
